+++ Extension/src/background/app/app-mv3.ts
@@ -51,6 +51,7 @@ import {
     PageStatsApi,
     UiApi,
     HitStatsApi,
+    iconsApi,
 } from '../api';
 import {
     UiService,
@@ -290,6 +291,9 @@ export class App {
         await engine.start();
 
         appContext.set(AppContextKey.IsInit, true);
+
+        // Update icons to hide "loading" icon
+        await iconsApi.update();
 
         await sendMessage({ type: MessageType.AppInitialized });