+++ Extension/src/pages/popup/components/AnimatedLoader/index.ts
@@ -16,4 +16,4 @@
  * along with AdGuard Browser Extension. If not, see <http://www.gnu.org/licenses/>.
  */
 
-export { SplashScreen } from './SplashScreen';
+export { AnimatedLoader } from './AnimatedLoader';